fix bug region cost

+++ src/main/java/gamein2/schedule/model/repository/RegionRepository.java
@@ -4,5 +4,5 @@ import gamein2.schedule.model.entity.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RegionRepository extends JpaRepository<Region,Long> {
-    Region findFirstByRegionId(Long regionId);
+    Region findFirstByRegionId(Integer regionId);
 }
